+++ pyproject.toml
@@ -4,14 +4,14 @@ build-backend = "hatchling.build"
 
 [project]
 name = "astra"
-version = "0.5.2"
+version = "0.6.0"
 authors = [{ name = "Korarei" }]
 description = "A build and deployment tool for AviUtl ExEdit2 scripts and plugins."
 readme = "README.md"
 license = "MIT"
 license-files = ["LICENSE"]
 requires-python = ">=3.11"
-dependencies = ["packaging>=23.0"]
+dependencies = ["packaging>=23.0", "filelock>=3.8.0", "pydantic>=2.0"]
 
 [project.scripts]
 astra = "astra.cli:main"
