+++ src/astra/_internal/schema.py
